% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_old.R
\name{.spadl_to_features}
\alias{.spadl_to_features}
\title{convert SPADL events to features}
\usage{
.spadl_to_features(events, fixtures_con = .settings$fixtures_con)
}
\arguments{
\item{events}{events from SPADL}

\item{fixtures_con}{fixtures db connection}
}
\value{
\code{tibble} representing features details.
}
\description{
convert SPADL events to features
}
