% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{.spadl_to_labels}
\alias{.spadl_to_labels}
\title{convert SPADL events to labels}
\usage{
.spadl_to_labels(events)
}
\arguments{
\item{events}{events from SPADL}

\item{fixtures_con}{fixtures db connection}
}
\value{
\code{tibble} representing labels details.
}
\description{
convert SPADL events to labels
}
