% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_db}
\alias{set_db}
\title{set up the package settings}
\usage{
set_db(
  json_config_file_name = "config_global_ra.json",
  json_config_path = "secrets",
  json_config_root = "",
  data_provider = c("opta", "inStat", "STATS", "whoScored"),
  database_type = c("localhost", "prod", "prod_backup"),
  feed_collections = c("events", "fixtures", "players", "gameEvents"),
  feats_collections = c("playerKeyPasses"),
  spadl_collections = c("features", "action_values"),
  spadl_config = "spadl_config.json",
  opta_config = "opta_config.json",
  scores_model_path = "model_scores.RDS",
  concedes_model_path = "model_concedes.RDS",
  project_name = "playeraction"
)
}
\arguments{
\item{json_config_file_name, json_config_path, json_config_root}{character info needed to access the global config file.}

\item{database_type}{character type of data-base connection}

\item{project_name}{character project-name}

\item{target_db}{character type of database playeraction}

\item{collections_provider}{character name of collection to connect to from
the \code{data_provider} family}

\item{collections_target}{character name of collection to connect to from
the \code{target_db} family}
}
\description{
set up the package settings
}
\details{
save the package settings in a top-level environment called
\code{.settings}. The function is called at loading time with default
arguments.
}
