% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{setDbAll}
\alias{setDbAll}
\title{set up the package settings}
\usage{
setDbAll(json_config_file_name = "config_global_ra.json",
  json_config_path = "secrets", json_config_root = "",
  target_db = "playeraction", database_type = c("local", "cloud"),
  collections_provider = c("events"),
  collections_target = c("spadl_actions"),
  project_name = "playeraction")
}
\arguments{
\item{json_config_file_name, json_config_path, json_config_root}{character info needed to access the global config file.}

\item{database_type}{character type of data-base connection}

\item{project_name}{character project-name}

\item{collections}{character name of collection to connect to from
the \code{data_provider} family}
}
\description{
set up the package settings
}
\details{
save the package settings in a top-level environment called
\code{.Settings}. The function is called at loading time with default
arguments.
}
